feat: change hardfrok countdown to use real date (#453)
