+++ src/pages/HardFork/index.tsx
@@ -13,7 +13,7 @@ import styles from './styles.module.scss'
 import glowingLine from './glowingLine.png'
 import { InfiniteMovingCard } from './InfiniteMovingCard'
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from '../../components/ui/Accordion'
-import { useCountdown } from '../../hooks'
+import { useCountdown, useEpochCountdown } from '../../hooks'
 import { explorerService, useStatistics } from '../../services/ExplorerService'
 import { ReactComponent as WarningCircle } from '../../assets/warning_circle.svg'
 import FlatCube from './FlatCube'
@@ -24,7 +24,8 @@ import { Link } from '../../components/Link'
 const targetVers = [0, 200] // 0.200.0
 
 export default function CountdownPage() {
-  const [days, hours, minutes, seconds, countdown] = useCountdown(ESTIMATED_ACTIVATION_TIME.end)
+  const { estimatedDate } = useEpochCountdown(ESTIMATED_ACTIVATION_TIME.epoch)
+  const [days, hours, minutes, seconds, countdown] = useCountdown(estimatedDate)
   const [isEnd, setIsEnd] = useState(false)
   const { t } = useTranslation()
 
@@ -63,7 +64,7 @@ export default function CountdownPage() {
     { label: 'target_epoch', value: ESTIMATED_ACTIVATION_TIME.epoch.toLocaleString('en') },
     {
       label: 'estimated_time',
-      value: dayjs(ESTIMATED_ACTIVATION_TIME.end).format('YYYY.MM.DD hh:mm:ss'),
+      value: dayjs(estimatedDate).format('YYYY.MM.DD hh:mm:ss'),
       tooltip: `UTC ${utcOffset > 0 ? `+ ${utcOffset}` : utcOffset}`,
     },
     { label: 'miner_percent', value: minerPercent },
@@ -194,7 +195,7 @@ export default function CountdownPage() {
                     Math.min(
                       100,
                       ((new Date().getTime() - ESTIMATED_ACTIVATION_TIME.start.getTime()) /
-                        (ESTIMATED_ACTIVATION_TIME.end.getTime() - ESTIMATED_ACTIVATION_TIME.start.getTime())) *
+                        (estimatedDate.getTime() - ESTIMATED_ACTIVATION_TIME.start.getTime())) *
                         100,
                     ),
                   )}%`,
